mp-3: finalized terrain and torus implementations.

+++ mp-3/shaders/torus-vertex.glsl
@@ -1,8 +1,7 @@
-precision mediump float;
+precision highp float;
 
 attribute vec4 a_vertexPosition;
 attribute vec3 a_vertexNormal;
-
 uniform mat4 u_modelViewMatrix;
 uniform mat4 u_projectionMatrix;
 
@@ -10,6 +9,5 @@ varying vec3 v_normal;
 
 void main() {
     v_normal = a_vertexNormal;
-
     gl_Position = u_projectionMatrix * u_modelViewMatrix * a_vertexPosition;
 }
